+++ casadocodigo-springboot/src/main/java/br/com/casadocodigo/loja/repository/UsuarioDAO.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.casadocodigo.loja.models.Usuario;
 
-public interface UsuarioDAO extends JpaRepository<Usuario, Long> {
+public interface UsuarioDAO extends JpaRepository<Usuario, Integer> {
 
 	Usuario findByEmail(String email);
